Refactor move DAME precompile to pallet module (#637)
* create precompile folder under pallet-asset-metadata-extender

* test works

* remove legacy folder

* remove old use of legacy folder

* fix clippy

* fmt
