+++ Cargo.toml
@@ -256,7 +256,6 @@ pallet-evm-evolution-collection = { path = "./precompile/evolution-collection", 
 pallet-evm-erc721 = { path = "./precompile/erc721", default-features = false }
 pallet-asset-metadata-extender = { path = "./pallets/asset-metadata-extender", default-features = false }
 pallet-evm-asset-metadata-extender = { path = "./precompile/asset-metadata-extender", default-features = false }
-pallet-evm-asset-metadata-extender-legacy = { path = "./precompile/asset-metadata-extender-legacy", default-features = false }
 pallet-parachain-staking = { path = "./pallets/parachain-staking", default-features = false }
 
 # Precompiles
