+++ pallets/asset-metadata-extender/Cargo.toml
@@ -18,10 +18,17 @@ sp-runtime = { workspace = true }
 sp-core = { workspace = true }
 frame-benchmarking = { optional = true, workspace = true }
 sp-std = { workspace = true }
+fp-evm = { workspace = true }
+precompile-utils = { workspace = true, default-features = false }
+sp-io = { workspace = true }
+pallet-evm = { workspace = true }
 
 [dev-dependencies]
-sp-io = { workspace = true }
 sp-runtime = { workspace = true }
+precompile-utils = { workspace = true, features = ["testing"] }
+pallet-balances = { workspace = true, features = ["std", "insecure_zero_ed"] }
+pallet-timestamp = { workspace = true }
+hex = { workspace = true }
 
 [features]
 default = ["std"]
@@ -34,6 +41,9 @@ std = [
 	"sp-runtime/std",
 	"frame-benchmarking/std",
 	"sp-std/std",
+	"fp-evm/std",
+	"precompile-utils/std",
+	"pallet-evm/std",
 ]
 
 runtime-benchmarks = [
