+++ pallets/laos-evolution/Cargo.toml
@@ -47,7 +47,7 @@ std = [
 	"precompile-utils/std",
 	"fp-evm/std",
 	"pallet-evm/std",
-	"laos-precompile-utils/std",
+	"laos-precompile-utils/std", # TODO remove?
 ]
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
