--- precompile/asset-metadata-extender-legacy/Cargo.toml
@@ -1,60 +1,0 @@
-[package]
-name = "pallet-evm-asset-metadata-extender-legacy"
-version = "0.1.0"
-edition = "2021"
-description = "Asset Metadata Extender precompile"
-repository = "https://github.com/freeverseio/laos"
-homepage.workspace = true
-authors.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-parity-scale-codec = { workspace = true, features = ["derive"] }
-scale-info = { workspace = true, features = ["derive"] }
-
-# Local pallet
-pallet-asset-metadata-extender = { workspace = true }
-
-# Local precompile utils
-laos-precompile-utils-macro = { workspace = true }
-laos-precompile-utils = { workspace = true }
-num_enum = { workspace = true }
-
-# Frontier
-fp-evm = { workspace = true, features = ["serde"] }
-pallet-evm = { workspace = true }
-
-# Frontier precompile utils
-precompile-utils = { workspace = true }
-
-# Substrate
-frame-support = { workspace = true }
-sp-core = { workspace = true }
-sp-runtime = { workspace = true }
-sp-std = { workspace = true }
-sp-io = { workspace = true }
-
-[dev-dependencies]
-hex = { workspace = true }
-frame-system = { workspace = true }
-parity-scale-codec = { workspace = true, features = ["derive"] }
-pallet-balances = { workspace = true, features = ["std", "insecure_zero_ed"] }
-pallet-timestamp = { workspace = true }
-precompile-utils = { workspace = true, features = ["testing"]}
-
-[features]
-default = ["std"]
-std = [
-	"fp-evm/std",
-	"pallet-evm/std",
-	"laos-precompile-utils/std",
-	"pallet-asset-metadata-extender/std",
-	"num_enum/std",
-	"frame-support/std",
-	"parity-scale-codec/std",
-	"scale-info/std",
-	"sp-core/std",
-	"sp-runtime/std",
-	"sp-std/std",
-]
