+++ precompile/asset-metadata-extender/src/mock.rs
@@ -15,14 +15,14 @@
 // along with LAOS.  If not, see <http://www.gnu.org/licenses/>.
 
 use core::str::FromStr;
+use fp_evm::{Precompile, PrecompileHandle};
 use frame_support::{
 	derive_impl, parameter_types, traits::FindAuthor, weights::constants::RocksDbWeight,
 };
-use precompile_utils::precompile_set::{AddressU64, PrecompileAt, PrecompileSetBuilder};
 use sp_core::{H160, U256};
 use sp_runtime::{traits::IdentityLookup, BuildStorage, ConsensusEngineId};
 
-use crate::{AssetMetadataExtenderPrecompile, AssetMetadataExtenderPrecompileCall};
+use crate::AssetMetadataExtenderPrecompile;
 
 // Configure a mock runtime to test the pallet.
 frame_support::construct_runtime!(
@@ -69,6 +69,7 @@ impl pallet_asset_metadata_extender::Config for Test {
 	type MaxUniversalLocationLength = MaxUniversalLocationLength;
 	type MaxTokenUriLength = MaxTokenUriLength;
 	type AccountIdToH160 = AccountIdToH160;
+	type GasWeightMapping = pallet_evm::FixedGasWeightMapping<Self>;
 }
 
 pub struct AccountIdToH160;
@@ -95,8 +96,8 @@ impl pallet_evm::Config for Test {
 	type AddressMapping = pallet_evm::IdentityAddressMapping;
 	type Currency = Balances;
 	type RuntimeEvent = RuntimeEvent;
-	type PrecompilesType = LaosPrecompiles<Self>;
-	type PrecompilesValue = PrecompilesInstance;
+	type PrecompilesType = MockPrecompileSet<Self>;
+	type PrecompilesValue = MockPrecompiles;
 	type ChainId = ();
 	type BlockGasLimit = BlockGasLimit;
 	type Runner = pallet_evm::runner::stack::Runner<Self>;
@@ -111,19 +112,12 @@ impl pallet_evm::Config for Test {
 pub const BLOCK_GAS_LIMIT: u64 = 15_000_000;
 pub const MAX_POV_SIZE: u64 = 5 * 1024 * 1024;
 
-pub type PrecompileCall = AssetMetadataExtenderPrecompileCall<Test>;
-
-pub type LaosPrecompiles<Test> = PrecompileSetBuilder<
-	Test,
-	(PrecompileAt<AddressU64<1>, AssetMetadataExtenderPrecompile<Test>>,),
->;
-
 frame_support::parameter_types! {
 	pub BlockGasLimit: U256 = U256::from(crate::mock::BLOCK_GAS_LIMIT);
 	pub const GasLimitPovSizeRatio: u64 = crate::mock::BLOCK_GAS_LIMIT.saturating_div(crate::mock::MAX_POV_SIZE);
 	/// 1 weight to 1 gas, for testing purposes
 	pub WeightPerGas: frame_support::weights::Weight = frame_support::weights::Weight::from_parts(1, 0);
-	pub PrecompilesInstance: LaosPrecompiles<Test> = LaosPrecompiles::new();
+	pub MockPrecompiles: MockPrecompileSet<Test> = MockPrecompileSet::<_>::new();
 }
 
 pub struct FindAuthorTruncated;
@@ -133,6 +127,33 @@ impl FindAuthor<H160> for FindAuthorTruncated {
 		I: 'a + IntoIterator<Item = (ConsensusEngineId, &'a [u8])>,
 	{
 		Some(H160::from_str("1234500000000000000000000000000000000000").unwrap())
+	}
+}
+
+#[derive(Default)]
+pub struct MockPrecompileSet<Test>(sp_std::marker::PhantomData<Test>);
+
+pub type MockAssetMetadataExtenderPrecompile = AssetMetadataExtenderPrecompile<Test>;
+
+impl<Test> MockPrecompileSet<Test>
+where
+	Test: pallet_evm::Config,
+{
+	pub fn new() -> Self {
+		Self(Default::default())
+	}
+}
+
+impl<Test> fp_evm::PrecompileSet for MockPrecompileSet<Test>
+where
+	Test: pallet_evm::Config + pallet_asset_metadata_extender::Config,
+{
+	fn execute(&self, handle: &mut impl PrecompileHandle) -> Option<fp_evm::PrecompileResult> {
+		Some(MockAssetMetadataExtenderPrecompile::execute(handle))
+	}
+
+	fn is_precompile(&self, _address: H160, _gas: u64) -> fp_evm::IsPrecompileResult {
+		fp_evm::IsPrecompileResult::Answer { is_precompile: true, extra_cost: 0 }
 	}
 }
