+++ runtime/klaos/Cargo.toml
@@ -56,7 +56,7 @@ pallet-laos-evolution = { workspace = true }
 
 pallet-evm-evolution-collection-factory = { workspace = true }
 pallet-evm-evolution-collection = { workspace = true }
-pallet-evm-asset-metadata-extender-legacy = { workspace = true }
+pallet-evm-asset-metadata-extender = { workspace = true }
 
 # Polkadot
 pallet-xcm = { workspace = true }
@@ -169,7 +169,7 @@ std = [
 	"pallet-ethereum/std",
 	"pallet-evm/std",
 	"pallet-evm-evolution-collection-factory/std",
-	"pallet-evm-asset-metadata-extender-legacy/std",
+	"pallet-evm-asset-metadata-extender/std",
 	"pallet-evm-evolution-collection/std",
 	"pallet-evm-chain-id/std",
 	"pallet-evm-precompile-modexp/std",
