+++ runtime/klaos/src/lib.rs
@@ -518,6 +518,7 @@ impl pallet_asset_metadata_extender::Config for Runtime {
 	type AccountIdToH160 = AccountIdToH160;
 	type MaxTokenUriLength = MaxTokenUriLength;
 	type MaxUniversalLocationLength = MaxUniversalLocationLength;
+	type GasWeightMapping = <Runtime as pallet_evm::Config>::GasWeightMapping;
 }
 
 impl pallet_sudo::Config for Runtime {
