+++ runtime/laos/src/precompiles/mod.rs
@@ -18,7 +18,7 @@
 
 use frame_support::parameter_types;
 
-use pallet_evm_asset_metadata_extender::AssetMetadataExtenderPrecompile;
+use pallet_asset_metadata_extender::precompiles::asset_metadata_extender::AssetMetadataExtenderPrecompile;
 use pallet_evm_precompile_blake2::Blake2F;
 use pallet_evm_precompile_bn128::{Bn128Add, Bn128Mul, Bn128Pairing};
 use pallet_evm_precompile_modexp::Modexp;
